changes due to review comments.
