+++ item_db.js
@@ -1,5 +1,3 @@
-/*eslint no-undef: "error"*/
-/*eslint no-console: ["error", { allow: ["warn", "error"] }] */
 /*eslint-env node*/
 const alasql = require('alasql');
