+++ tests/items.test.js
@@ -1,11 +1,26 @@
-/*eslint no-undef: "error"*/
 /*global describe*/
 /*global test*/
-/*eslint no-console: ["error", { allow: ["warn", "error"] }] */
 /*eslint-env node*/
+const app = require('../app.js');
 const request = require('supertest');
-const app =  require('../app.js');
 var cookie;
+
+describe('POST /login/', function() {
+	test('POST /login/ respond with json', function(done) {
+		request(app)
+			.post('/login/')
+			.send({'username':'Screwee','password':'Password01'})
+			.set('Content-Type', 'application/json')
+			.set('Accept', 'application/json')
+			.expect(200)
+			.expect('Hello user Screwee')
+			.end(function(err,res) {
+				if (err) return done(err);
+				cookie = res.headers['set-cookie'];
+				done();
+			});
+	});
+});
 
 describe('POST /items', function() {
 	test('POST /items respond with json', function(done) {
@@ -50,22 +65,6 @@ describe('POST /items', function() {
 	});
 });
 
-describe('POST /login/', function() {
-	test('POST /login/ respond with json', function(done) {
-		request(app)
-			.post('/login/')
-			.send({'username':'Screwee','password':'Password01'})
-			.set('Content-Type', 'application/json')
-			.set('Accept', 'application/json')
-			.expect(200)
-			.expect('Hello user Screwee')
-			.end(function(err,res) {
-				if (err) return done(err);
-				cookie = res.headers['set-cookie'];
-				done();
-			});
-	});
-});
 
 describe('GET /items', function() {
 	test('GET respond with json', function(done) {
@@ -180,71 +179,6 @@ describe('DELETE /items/', function() {
 			.set('Accept', 'application/json')
 			.set('cookie', cookie)
 			.expect(204)
-			.end(function(err) {
-				if (err) return done(err);
-				done();
-			});
-	});
-});
-
-describe('User is not logged in', function() {
-	test('GET /items', function(done) {
-		request(app)
-			.get('/items')
-			.set('Accept', 'application/json')
-			.expect(403, 'You are not logged in')
-			.end(function(err) {
-				if (err) return done(err);
-				done();
-			});
-	});
-	test('DELETE /items/', function(done) {
-		request(app)
-			.delete('/items')
-			.set('Accept', 'application/json')
-			.expect(403, 'You are not logged in')
-			.end(function(err) {
-				if (err) return done(err);
-				done();
-			});
-	});
-});
-
-describe('GET /profile/', function() {
-	test('GET /profile/ returns users Information', function(done) {
-		request(app)
-			.get('/profile/')
-			.set('Accept', 'application/json')
-			.set('cookie', cookie)
-			.expect(200, [{
-				id: 1,
-				username: 'Screwee'
-			}])
-			.expect('Content-Type',/json/)
-			.end(function(err) {
-				if (err) return done(err);
-				done();
-			});
-	});
-});
-describe('GET /logout/', function() {
-	test('GET /logout/ logout user from session', function(done) {
-		request(app)
-			.get('/logout/')
-			.set('Accept', 'application/json')
-			.set('cookie', cookie)
-			.expect(200, 'User logged out')
-			.end(function(err) {
-				if (err) return done(err);
-				done();
-			});
-	});
-	test('GET /profile/ returns that user is not logged in', function(done) {
-		request(app)
-			.get('/profile/')
-			.set('Accept', 'application/json')
-			.set('cookie', cookie)
-			.expect(403, 'You are not logged in')
 			.end(function(err) {
 				if (err) return done(err);
 				done();
